Split lib.rs into submodules
Doesn't delete any code right now
