+++ src/manually_traced/core.rs
@@ -9,7 +9,6 @@ use core::marker::PhantomData;
 use core::num::Wrapping;
 
 use crate::prelude::*;
-use crate::GcDirectBarrier;
 
 use zerogc_derive::unsafe_gc_impl;
